Add new module using check status
